Coin::_21() - 100%

+++ src/AAA.hpp
@@ -90,7 +90,7 @@ extern Heap* data_0208b720;
 void func_02020354(i8);
 void func_02012398(i32, Vec3_32*);
 bool func_020202a0();
-i32 func_0202040c(i8);
+i32 func_0202040c(i32);
 
 //
 bool func_0204d82c();
